fix: dist

+++ apps/web/scripts/postbuild.sh
@@ -3,14 +3,14 @@
 # 
 # change dir structure
 # ./dist/{html} -> ./dist/2024/{html}
-# 
-# if [ ! -d "./.output" ]; then
-#     echo "can not find .output"
-#     exit 1;
-# fi
+mkdir ./dist
+if [ ! -d "./.nuxt/dist" ]; then
+    echo "can not find .nuxt/dist"
+    exit 1;
+fi
 
 mkdir ./tmp
-mv ./.output/* ./tmp
+mv ./.nuxt/dist/* ./tmp
 mkdir ./dist/2024
 mv ./tmp/* ./dist/2024
 rm -rf ./tmp
